fix types in functions and add instruction cycle in main, need to individually test if instructions are being carried out correctly - ibm chip8 logo test fails

+++ include/display.h
@@ -18,7 +18,7 @@ void print_array();
 void init_display();
 
 // write to display buffer
-void wt_d_buffer(uint16_t *memory, uint8_t *VF, int VX, int VY);
+void wt_d_buffer(uint8_t *memory, uint8_t *VF, int VX, int VY);
 
 // draw on SDL screen using renderer
 void draw(SDL_Renderer *renderer);
